backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.1
httpx==0.27.2
python-dotenv==1.0.1
python-multipart==0.0.9
aiofiles==24.1.0
pydantic==2.12.3
requests==2.32.3
pyyaml==6.0.2
psutil==6.1.0
docker==7.1.0
pydantic-ai-slim[openai]==1.78.0
holidays
